BACKEND/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
matplotlib==3.8.2
Pillow==10.2.0
Werkzeug==3.0.1